requirements.txt

coreapi==2.3.3
Django==2.0.7
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.8.2